Importing stock data:
Runs PE algorithms on stock data imported from Yahoo Finance, and uses an adaptive threshold method to make a decision as to when action should be taken on a stock (open/close a short or long position). Further extensions to the code are planned to categories what directions these actions should take, e.g. whether a recommended investment should be a long or a short investment.